Do not make separate /boot partition
There is no need in it with supported bootloaders (Grub2 only) on x86 and ARM.
Avoid running out of free space in /boot if many kernels are installed.

Conan-Kudo (ngompa@) noted that a separate /boot partition will be useful
if we want to add (BTRFS) snapshots to Grub's menu because there is no
other way to list snapshots than store a specification of them inside /boot
when creating a snapshot.
https://github.com/snapshotmanager/boom may be used in the future
and then we may restore separate /boot.