Make network setup work always
I do not see sense in asking users to use DE's tools to edit network configuration when running inside a LiveCD,
let's just enable in-Anaconda configuration of network always.

It interact with system NetworkManager, the same that DE's GUI tools interact with, but what if NetworkManager is not used?
Let's just ignore such cases for now.