Remove --ignore-broken test (#test)
In the past the --ignore-broken was disabled on RHEL by Anaconda configuration
file, however, in the tests the RHEL conf is not loaded so the test still worked
here.

After change b24574113fce6895e32af5f8171e732a3dab134f in pykickstart this
functionality was moved to the pykickstart project so it's not allowed to be
used on RHEL anymore.

Anaconda code should work as expected because we are disabling this anyhow by
the configuration but we have to remove the test to fix our tests right now.