driver_updates: add lot of debug logs for easier debugging (#2154906)
Without these it could be real pain to solve bugs here. Most of the time special
HW is required to be able to reproduce this and even with the HW there is not
100% sure that we are able to reproduce the issue.

These logs will always go to journal log and with 'inst.debug' or 'rd.debug' on
command line they will show up also in the console during boot. That is great
help in case the code will freeze and boot process blocked.

Also fix test scanning files for mount keyword. It is now used in the debug logs.

Related: rhbz#2154906