Show an error message when subscription token transfer fails
Instead of a hard failure and confusing traceback show an error
message when subscription token transfer fails.

The main expected trigger for this happening is users with an expired
free Red Hat developer account, so we mention that in the error message.

Also if we actually get to the token transfer attempt, there must be
a working installation source and the system will be most likely functional.
It will just not have a subscription - so make the error non-fatal
and make it possible to continue with the installation.

Resolves: rhbz#2057472