Extend LiveCD capabilities
LiveCD can be rebootedЖ
We in ROSA run Anaconda in liveinst (rsync) mode in both desktop and server ISOs,
both in GUI and TUI modes, system must be rebootable by a button in Anaconda in all these cases.

auditd and rsyslogd can also be managed inside a LiveCD.

A random hostname is generated by anaconda-livecd-init.service, copy it into the installed system.