Implement EULA spoke based on initial-setup
Copied code from initial-setup and adapted it for Anaconda.

It makes sense to show EULA before installation, not just text
that it is in some file after the installation.

TODO: It would be great to automatically choose license in the current language.
Currently it is worked around in ROSA by creating a drop-in config:
https://abf.io/import/anaconda/blob/6dc5b78efd/anaconda.sh#lc-77

Closes: https://github.com/rhinstaller/anaconda/discussions/4686

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>