Stop infinite wait for CDROM when KS is processed (#2209599)
Recently we introduced startup wait about 5 secs to give OEMDRV disks
posibility to show up before we switch out from initrd. It was
implemented by rhbz#1770969.

This unfortunately conflicts with our older feature implemented where we
wait infinitely on CDROM device if kickstart is specified to come from a
disk implemented by rhbz#1168902.

What happens is that waiting for OEMDRV checks the `main_loop` variable
from dracut which is increased when Dracut even loop cycle without new
input. However, the infinite wait on kickstart on CDROM resets this
value to avoid dracut timeout.

The fix will avoid reseting the `main_loop` variable if the kickstart is
already processed. Drawback of this solution is that we will wait
another 5 seconds before moving the processing further but because we
are not sure how many times the event loop iterated before this seems
like the most robust solution.

Resolves: rhbz#2209599